% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{blockm}
\alias{blockm}
\title{Construct a block-diagonal matrix}
\usage{
blockm(A, B, C)
}
\arguments{
\item{A}{Numerical matrix. Upper diagonal block.}

\item{B}{Numerical matrix. Off diagonal block.}

\item{C}{Numerical matrix. Lower diagonal block}
}
\value{
A block diagonal matrix of the form \eqn{\begin{bmatrix} A & B \\ B' & C \end{bmatrix}}.
}
\description{
Construct a block-diagonal matrix
}
\details{
This function is typically called for constructing the covariance matrix of $(x,y)$.
}
