% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{bootstrap_inner.po2m}
\alias{bootstrap_inner.po2m}
\title{Calculate standard errors for the inner relation coefficient B}
\usage{
bootstrap_inner.po2m(fit, X, Y, rep.cores = 1, rep.K = 5, ...)
}
\arguments{
\item{fit}{A PO2PLS fit of class po2m}

\item{X}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{rep.cores}{Positive integer. Number of cores.}

\item{rep.K}{Positive integer. Number of repeats.}

\item{...}{Additional arguments for the PO2PLS fit. In particular, one may specify \code{steps=100, init_param=fit$par, verbose=FALSE}}
}
\value{
A vector with the standard errors for B per component
}
\description{
Calculate standard errors for the inner relation coefficient B
}
