% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{E_step_slow}
\alias{E_step_slow}
\title{Expectation step (slower version)}
\usage{
E_step_slow(X, Y, params)
}
\arguments{
\item{X}{Data matrix}

\item{Y}{Data matrix}

\item{params}{List with parameters}
}
\description{
Internal function only
}
\keyword{internal}
