% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{E_step_test}
\alias{E_step_test}
\title{Expectation step}
\usage{
E_step_test(
  dataXY,
  p,
  q,
  r,
  rx,
  ry,
  N,
  SigmaZ,
  GammaEF,
  invZtilde,
  Gamma,
  GGef,
  EZc,
  Szz,
  W,
  C,
  Wo,
  Co,
  SigT,
  SigH,
  sig2E,
  sig2F
)
}
\description{
Internal function only
}
\keyword{internal}
