% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{generate_data}
\alias{generate_data}
\title{Generate two datasets based on PO2PLS}
\usage{
generate_data(N, params, distr = rnorm)
}
\arguments{
\item{N}{Positive integer. Sample size to be simulated}

\item{params}{A list as generated by \code{\link{generate_params}}}

\item{distr}{Function. One of the random number generator functions (e.g. rnorm) for generating latent variables}
}
\value{
A list containing the X and Y matrix
}
\description{
Generate two datasets based on PO2PLS
}
