% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{jitter_params}
\alias{jitter_params}
\title{Jitter PO2PLS parameters}
\usage{
jitter_params(params, amount = NULL)
}
\arguments{
\item{params}{List with parameters}

\item{amount}{Amount of jitter}
}
\description{
To be rewritten
}
\keyword{internal}
