% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{Lemma}
\alias{Lemma}
\title{Implements block-wise inverse}
\usage{
Lemma(X, SigmaZ, invZtilde, Gamma, sig2E, sig2F, p, q, r, rx, ry)
}
\arguments{
\item{X}{Concatenated data matrix}

\item{SigmaZ}{Variance of Z}

\item{invZtilde}{Inverse of Z_tilde}

\item{Gamma}{Gamma}

\item{sig2E}{Variance of E}

\item{sig2F}{Variance of F}

\item{p}{X dimensions}

\item{q}{Y dimensions}

\item{r}{Positive integer. Number of joint PLS components. Must be positive!}

\item{rx}{Non-negative integer. Number of orthogonal components in \eqn{X}. Can be 0}

\item{ry}{Non-negative integer. Number of orthogonal components in \eqn{Y}. Can be 0}
}
\description{
Internal use only
}
\keyword{internal}
