% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_output.R
\name{LRT}
\alias{LRT}
\alias{LRT.po2m}
\title{Likelihood ratio test for two nested PO2PLS fits}
\usage{
LRT(fit, fit0, digits = 3, ...)

\method{LRT}{po2m}(fit, fit0, digits = 3, ...)
}
\arguments{
\item{fit}{A PO2PLS fit of class po2m}

\item{fit0}{A second PO2PLS fit of class po2m}

\item{digits}{Number of decimals to output}

\item{...}{For consistency}
}
\value{
The chi-square statistics (2 times the difference in log-likelihood), the degrees of freedom and the p-value of the comparison.
}
\description{
Perform a likelihood ratio test to compare two nested fits from a PO2PLS model.
}
