% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{M_step}
\alias{M_step}
\title{Maximization step}
\usage{
M_step(E_fit, params, X, Y, orth_type = c("SVD", "QR"))
}
\arguments{
\item{E_fit}{An object returned by E_step}

\item{params}{List with parameters}

\item{X}{Data matrix}

\item{Y}{Data matrix}

\item{orth_type}{Character. One of "SVD" or "QR". Best left set to "SVD"}
}
\description{
Internal use only
}
\keyword{internal}
