% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\docType{package}
\name{PO2PLS-package}
\alias{PO2PLS-package}
\title{PO2PLS: Probabilistic Two-way Orthogonal Partial Least Squares}
\description{
This package implements the probabilistic O2PLS method.
}
\author{
Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}),
Jeanine Houwing-Duistermaat (\email{J.J.Houwing@lumc.nl}),
Geurt Jongbloed (\email{G.Jongbloed@tudelft.nl}),
Szymon Kielbasa (\email{S.M.Kielbasa@lumc.nl}),
Hae-Won Uh (\email{H.Uh@lumc.nl}).

Maintainer: Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}).
}
\keyword{Probabilistic-O2PLS}
