% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{PO2PLS}
\alias{PO2PLS}
\title{Perform O2-PLS with two-way orthogonal corrections}
\usage{
PO2PLS(X, Y, r, rx, ry, steps = 1e+05, tol = 1e-06,
  init_param = "o2m", orth_type = "SVD", random_restart = FALSE,
  homogen_joint = FALSE, null_B = FALSE)
}
\arguments{
\item{X}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{r}{Positive integer. Number of joint PLS components. Must be positive!}

\item{rx}{Non-negative integer. Number of orthogonal components in \eqn{X}. Can be 0}

\item{ry}{Non-negative integer. Number of orthogonal components in \eqn{Y}. Can be 0}

\item{steps}{Positive integer. Number of EM steps to perform}

\item{tol}{Positive double. Tolerance of deciding if the likelihood increment is small enough to conclude convergence.}

\item{init_param}{Character. Should be one of "random", "o2m" or "unit". Specifies which kind of parameters should be generated.}

\item{orth_type}{Character. One of "SVD" or "QR". Best left set to "SVD"}

\item{random_restart}{Not to be used}

\item{homogen_joint}{Boolean. Should U=T be assumed? Best left to statistical expert.}
}
\value{
A list with
\describe{
\item{parameters}{Estimated PO2PLS parameters}
\item{latent_vars}{Conditional expectation and variances of latent variables}
\item{meta_data}{Meta data to be used for \code{print} and \code{summary}}
}
}
\description{
NOTE THAT THIS FUNCTION DOES NOT CENTER NOR SCALES THE MATRICES! Any normalization you will have to do yourself.
It is best practice to at least center the variables though.
}
