% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_output.R
\name{PO2PLS_to_po2m}
\alias{PO2PLS_to_po2m}
\title{Convert old PO2PLS fit to new-style PO2PLS fit}
\usage{
PO2PLS_to_po2m(fit, X, Y)
}
\arguments{
\item{fit}{A PO2PLS of old style}

\item{X}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}
}
\description{
Internal use only
}
\keyword{internal}
