% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_output.R
\name{print.loglik.po2m}
\alias{print.loglik.po2m}
\title{Prints log-likelihood statistics of a PO2PLS fit}
\usage{
\method{print}{loglik.po2m}(x, digits = 3, ...)
}
\arguments{
\item{x}{A log-likelihood object of class loglik.po2m, produced within a PO2PLS fit}

\item{digits}{Number of decimals to output}

\item{...}{For consistency}
}
\description{
Prints log-likelihood statistics of a PO2PLS fit
}
