% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_output.R
\name{print.po2m}
\alias{print.po2m}
\title{Prints an overview of a PO2PLS fit}
\usage{
\method{print}{po2m}(x, digits = 3, ...)
}
\arguments{
\item{x}{A PO2PLS fit of class po2m}

\item{digits}{Number of decimals to output}

\item{...}{For consistency}
}
\description{
Prints an overview of a PO2PLS fit
}
