% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_output.R
\name{print.summary.po2m}
\alias{print.summary.po2m}
\title{Prints a summary of a PO2PLS fit}
\usage{
\method{print}{summary.po2m}(x, digits = 3, ...)
}
\arguments{
\item{x}{A PO2PLS fit of class po2m}

\item{digits}{Number of decimals to output}

\item{...}{For consistency}
}
\description{
Prints a summary of a PO2PLS fit
}
