% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_output.R
\name{summary.po2m}
\alias{summary.po2m}
\title{Calculates a summary of a PO2PLS fit}
\usage{
\method{summary}{po2m}(object, digits = 3, ...)
}
\arguments{
\item{object}{A PO2PLS fit of class po2m}

\item{digits}{Number of decimals to output}

\item{...}{For consistency}
}
\description{
Calculates a summary of a PO2PLS fit
}
