% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{variances.PO2PLS}
\alias{variances.PO2PLS}
\title{Calculate the variance covariance matrix of the estimated PO2PLS parameters}
\usage{
variances.PO2PLS(fit, data, type_var = c("complete", "component",
  "variable"))
}
\arguments{
\item{fit}{A PO2PLS fit of class po2m}

\item{data}{The datasets used in the fit}

\item{type_var}{String. Type of covariance matrix sought}
}
\value{
A covariance matrix and standard errors
}
\description{
Calculate the variance covariance matrix of the estimated PO2PLS parameters
}
