% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\docType{package}
\name{PO2PLS-package}
\alias{PO2PLS-package}
\title{PO2PLS: Probabilistic Two-way Orthogonal Partial Least Squares}
\description{
This package implements the probabilistic O2PLS method.
}
\author{
Said el Bouhaddani,
Jeanine Houwing-Duistermaat,
Geurt Jongbloed,
Hae-Won Uh.

Maintainer: Said el Bouhaddani (\email{s.el_bouhaddani@outlook.com}).
}
\keyword{Probabilistic-O2PLS}
