% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO2PLS_functions.R
\name{variances_inner.po2m}
\alias{variances_inner.po2m}
\title{Calculate standard errors for the inner relation coefficient B}
\usage{
variances_inner.po2m(fit, X, Y)
}
\arguments{
\item{fit}{A PO2PLS fit of class po2m}

\item{X}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}

\item{Y}{Numeric matrix. Other types will be coerced to matrix with \code{as.matrix} (if this is possible)}
}
\value{
A vector with the standard errors for B per component
}
\description{
Calculate standard errors for the inner relation coefficient B
}
